Send the stop_date in server timezone.
Encapsulate the html showing the server's timezone into
a `span` with id and data attribute.

When clicking the OFF button, read the value from there
and use `moment-with-timezone` to converr the date which is
in the client's timezone to the server's timezone.

For the purpose we need to vendor the `moment-timezone-with-data.js`,
because it contains all the data about timezones, time differences, etc.

Without doing that, all convertions fall back to UTC and I found no other way
to download the file (via npm or something else).